Add basic script to query PGS api for pubs
This is a very basic script so far; all it does it read in a csv
file that contains PMIDs in a column, queries the API for each PMID,
and then prints the response. It also counts the number of responses
that have a result to get the total count of PMIDs that have a PGP
record, and prints it out.