Rework PGS querying script to use new client SDK
This new script searches the PGS catalog by pmid for both scores and
metrics. It also saves the output as json results as returned by the
API, instead of creating a data frame that can be plotted by the
report. This way, we have the full API record for all the scores and
metrics saved.