Query for pubs and then PGS associated with those pubs
The previous set of queries did not return the "evaluation" PGS
associated with a publication. To fix this, first query for all
publication records in PGS catalog associated with all of the PRIMED
PubmedIDs. Then compile the full set of PGS associated with those
publications. Next, pull all information about all the PGS ids from
the API (instead of searching for PGS by pmid). Finally, pull all
of the metrics associated with each pmid. This also means storing
the pubs results in a text file.