Add links to some tables
Link to the PGP_ID and the PGS_ID in some of the tables. For some
reason, trying to link in the table that has grouped rows does not
work - it splits the URL into the columns instead of showing the
percentages - so there is no link there. Not sure why this is happening.