Make trait display in tables nicer
Use the trait_efo mapped information instead of the trait_reported
information. Add a link to the EFO detail page in PGS catalog. Under
the hood, add a function that pulls EFO trait information out of
the record for a score and adds it to the data frame. This way, we
can reuse code without having to rewrite it.