Use a for loop to retrieve pgs records
Trying to retrieve all scores filtered to only the requested pgs_ids
ended up giving a paginated result, and it was unclear how to handle
that in the swagger-codegen framework. (I'm sure there's a way other
than making a get request to response.next and then converting the
results to Score objects...) To deal with this and for consistency
with other functions, use a for loop instead.