Add starter code to match a set of variants to pgs catalog scores
This code is a framework for how we might do the queries, but it
looks like there is already soem code for matching variants to
pgs catalog scores:
https://github.com/PGScatalog/pygscatalog

so possibly better to just use that code instead of writing our own.
Commit this file so we have a record of it anyway.