Add comments about PGS score ids that are currently broken
The pgscatalog-combine command fails to run on ~15 ids from the catalog.
There is currently an issue and PR to fix this, so until that is
done I'm just adding the list that are currently failing. Once
the PR is merged and a new version of pygscatalog-utils is released,
I will update to that version.

Issue: https://github.com/PGScatalog/pygscatalog/issues/44