Correctly reference stylesheet
When I restructured the repostiory in a previous branch, the path to
the stylesheet was no longer correct. The workflow still worked, since
the docker image in the WDL hadn't been updated. Now that we're updating
the docker image, we also have to update the Rmd to have the correct
path to the stylesheet.